Refactored TO IT to use common base class
